+++ CHANGELOG.md
@@ -7,22 +7,33 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Add tree-sitter functionality for extracting expressions from source code
+
 ## [0.2.0] - 2021-10-12
+
 ### Changed
+
 - Upgrade to `@cucumber/cucumber-expressions 14.0.0`
 
 ## [0.1.0] - 2021-09-07
+
 ### Added
+
 - Document Formatting
-([#1732](https://github.com/cucumber/common/pull/1732)
-[aslakhellesoy](https://github.com/aslakhellesoy))
+  ([#1732](https://github.com/cucumber/common/pull/1732)
+  [aslakhellesoy](https://github.com/aslakhellesoy))
 
 ## [0.0.1] - 2021-09-02
+
 ### Added
+
 - First release
 
-[Unreleased]: https://github.com/cucumber/language-server/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/cucumber/language-server/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/cucumber/language-server/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/cucumber/language-server/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/cucumber/common/tree/v0.0.1
+
 ellesoy